Python3.6/requirements.txt
absl-py==0.7.1
astor==0.7.1
backports.csv==1.0.7
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.0
cycler==0.10.0
feedparser==5.2.1
future==0.17.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
jaraco.functools==2.0
jieba==0.39
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lxml==4.3.2
Markdown==3.0.1
matplotlib==3.0.3
mock==2.0.0
more-itertools==6.0.0
mysqlclient==1.4.2.post1
nltk==3.4
numpy==1.16.2
pandas==0.24.2
Pattern==3.6
pbr==5.1.3
pdfminer.six==20181108
Pillow==5.4.1
pip==10.0.1
portend==2.3
protobuf==3.7.0
pycryptodome==3.8.0
pyparsing==2.3.1
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2018.9
pywin32==224
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
setuptools==39.0.1
singledispatch==3.4.0.3
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.8
tempora==1.14
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Theano==1.0.4
torch==1.0.1
torchvision==0.2.2.post3
urllib3==1.24.1
Werkzeug==0.15.1
wheel==0.33.1
zc.lockfile==1.4
